Move icons route to the respective app
